buffer bug fix and collectd handler bug fix

+++ plugins/handler/collectd-metrics/pkg/collectd/collectd.go
@@ -15,7 +15,7 @@ type Metric struct {
 	Dstypes        []string     `json:"dstypes"`
 	Dsnames        []string     `json:"dsnames,omitempty"`
 	Time           cdtime.Time  `json:"time"`
-	Interval       int64        `json:"interval"`
+	Interval       float64      `json:"interval"`
 	Host           string       `json:"host"`
 	Plugin         string       `json:"plugin"`
 	PluginInstance string       `json:"plugin_instance,omitempty"`
